Fix methyl extraction (#95)
* Memory and extraction fix
* Edge case improved for extraction
* Long runtime and nih_s10 partition default parameter
* Improve if statements and file transfers and add more if statements
* Fix indexing file name error
* Set methyl extraction to be paired end bams
* Better echo formatting of logs
* Prevent rm error if file not present
* Better loop formation in logs
* Fix sorting error

+++ Methylseq/deduplicate.sh
@@ -1,8 +1,9 @@
 #!/bin/bash
 
 #script to deduplicate the output from mapping.
-
+echo ""
 echo "entering deduplicate script"
+echo ""
 
 dedup_input=$1
 dedup_output=$2
@@ -38,16 +39,38 @@ module load bismark/0.22.3
 	echo "Deduplication requested for mapping output $dedup_input"
 	echo "Expected deduplication output file is $dedup_output"
 
-	if [ ! -f "$dedup_output" ] || [ ! -f "$dedup_report" ]; then
+#copy over output files if they are already made
+dedup_output_name=$(basename "${dedup_output}")
+dedup_report_name=$(basename "${dedup_report}")
+
+dedup_input_name=$(basename "${dedup_input}")
+
+
+
+#if there is not the deduplicate output and deduplicate report in the output directory, then transfer files needed to make it and make it. else skip
+	if [ ! -f "$output_directory/$dedup_output_name" ] || [ ! -f "$output_directory/$dedup_report_name" ]; then
+		#transfer files needed if not present yet in temp directory.
+		if [ ! -f "$dedup_input" ]; then
+			rsync -vur --include="${dedup_input_name}" --exclude="*" "$output_directory/" $output_temp_directory
+		fi
+
 		echo "Begin deduplicating $dedup_input"
             	deduplicate_bismark -p --bam $dedup_input --output_dir $output_temp_directory -o $dedup_input 
             	#-o uses $dedup_input as output file basename because bismark modifies it to add deduplicated.bam ... -o just wants the basename to use for the output file.
             	#output_dir needed, otherwise outputs get written to the working directory!
-            	echo "Finished deduplicating $dedup_input"
-    fi
+        echo ""
+        echo "Finished deduplicating $dedup_input"
+        echo ""
 
-#checkpoint
-rsync -vur $output_temp_directory/ $output_directory
+        #checkpoint
+		rsync -vur $output_temp_directory/ $output_directory
+
+	else
+	echo ""
+	echo "$(basename "$dedup_input") has already been deduplicated"
+	echo ""
+
+    fi
 
 #TO DO: This script could use testing code that tests each 'then' and 'else' sections of each 'if' statement! So a total of 4 test cases.
